render

This function create a `Rmarkdown` script and render it as a HTML document. The HTML document is a quality-control report using all the metrics from `visualise()` based on recommendations from NanoString.

render(nacho_object, colour = "CartridgeID",
  output_file = "NACHO_QC.html", output_dir = getwd(),
  show_legend = TRUE, keep_rmd = FALSE)

Arguments

nacho_object
[list] List obtained from summarise or normalise.

colour
[character] Character string of the column in ssheet_csv or more generally in nacho_object$nacho to be used as grouping colour.

output_file
[character] The name of the output file. If using `NULL` then the output filename will be based on filename for the input file. If a filename is provided, a path to the output file can also be provided. Note that the `output_dir` option allows for specifying the output file path as well, however, if also specifying the path, the directory must exist. If `output_file` is specified but does not have a file extension, an extension will be automatically added according to the output format. To avoid the automatic file extension, put the output_file value in `I()`, e.g., I('my-output').

output_dir
[character] The output directory for the rendered output_file. This allows for a choice of an alternate directory to which the output file should be written (the default output directory is the working directory, i.e., `getwd()`). If a path is provided with a filename in `output_file` the directory specified here will take precedence. Please note that any directory path provided will create any necessary directories if they do not exist.

show_legend
[logical] Boolean to indicate whether the plot legends should be plotted (TRUE) or not (FALSE). Default is TRUE.

keep_rmd
[logical] Boolean to indicate whether the Rmd file used to produce the HTML report is copied to the directory provided in `output_dir`. Default is FALSE.

Examples

data(GSE74821)
render(nacho_object = GSE74821, output_dir = NULL)

#> 
#> 
#> processing file: file65633b4d0f56.Rmd

#>   |                                                                         |                                                                 |   0%  |                                                                         |...........                                                      |  17%
#>   ordinary text without R code
#> 
#>   |                                                                         |......................                                           |  33%
#> label: setup (with options) 
#> List of 1
#>  $ include: logi FALSE
#> 
#>   |                                                                         |................................                                 |  50%
#>   ordinary text without R code
#> 
#>   |                                                                         |...........................................                      |  67%
#> label: nacho_qc

#> Warning: Transformation introduced infinite values in continuous y-axis

#> Warning: Transformation introduced infinite values in continuous y-axis

#>   |                                                                         |......................................................           |  83%
#>   ordinary text without R code
#> 
#>   |                                                                         |.................................................................| 100%
#> label: session_info (with options) 
#> List of 1
#>  $ results: chr "markup"
#> 
#> 

#> output file: file65633b4d0f56.knit.md

#> /usr/bin/pandoc +RTS -K512m -RTS file65633b4d0f56.utf8.md --to html4 --from markdown+autolink_bare_uris+ascii_identifiers+tex_math_single_backslash+smart --output /tmp/RtmpB16ppG/NACHO_QC.html --email-obfuscation none --self-contained --standalone --section-divs --table-of-contents --toc-depth 2 --variable toc_float=1 --variable toc_selectors=h1,h2 --variable toc_smooth_scroll=1 --variable toc_print=1 --template /home/travis/R/Library/rmarkdown/rmd/h/default.html --no-highlight --variable highlightjs=1 --number-sections --variable 'theme:simplex' --include-in-header /tmp/RtmpB16ppG/rmarkdown-str6563154a3be.html --mathjax --variable 'mathjax-url:https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' 

#> Error: pandoc document conversion failed with error 99